requirements.txt
click==8.1.3
distlib==0.3.5
dnspython==2.2.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
pymongo==4.2.0
python-dotenv==0.20.0
six==1.16.0
virtualenv==20.16.3
Werkzeug==2.2.2
zipp==3.8.1